Moved query and quote into sqlite
